ARMv7-A: full aster-kernel lib compiles for armv7a-none-eabi
Complete the kernel-side ARM support so the entire kernel crate (ostd + all
components + kernel logic) builds for armv7a-none-eabi:

- kernel arch/arm layer (LinuxAbi, SigContext, ToFaultSignal) wired into all
  the signal/vdso/ELF cfg gates (signal/mod.rs, c_types.rs, rt_sigreturn.rs,
  thread/exception.rs, elf_file.rs [32-bit ELF class + Machine::Arm],
  process_vm vdso_base, load_elf, sig_disposition).
- ARMv7 vDSO: hand-written, C-free vdso_arm.so (tools/build_vdso_arm.sh),
  __VDSO_RT_SIGRETURN_OFFSET=0x200, VDSO_VMO_LAYOUT.
- Component drivers: PL011 UART + ECAM PCI arch modules, virtio 64-bit
  DMA-address/capacity split fixed for 32-bit; excluded aster-mlsdisk from
  the arm build (pervasively 64-bit, not on the boot path).
- Misc 32-bit portability: futex/exfat/overlayfs/uts_ns arithmetic and
  sentinel-literal fixes.

Next: link a bootable image via cargo-osdk and boot on qemu-system-arm.

+++ kernel/comps/virtio/src/device/block/mod.rs
@@ -191,11 +191,13 @@ impl ConfigManager<VirtioBlockConfig> {
     pub(self) fn capacity_sectors(&self) -> usize {
         let cap_low = self
             .read_once::<u32>(offset_of!(VirtioBlockConfig, capacity))
-            .unwrap() as usize;
+            .unwrap() as u64;
         let cap_high = self
             .read_once::<u32>(offset_of!(VirtioBlockConfig, capacity) + 4)
-            .unwrap() as usize;
+            .unwrap() as u64;
 
-        (cap_high << 32) | cap_low
+        // The device capacity is a 64-bit sector count; compute in `u64` and
+        // narrow to `usize` (on 32-bit targets this bounds addressable disks).
+        (((cap_high << 32) | cap_low) as usize)
     }
 }
